+++ CHANGELOG.md
@@ -1,11 +1,19 @@
-## 2.0.0 (2021-03-20) REFACTOR
+## 2.0.1 (2021-04-26) NEXT
+
+### Added Features
+
+- add `checkConnectionsConsistency` method
+- add `import` & `export` JSON Listeners
+- update README.md & APIHook.md
+
+## 2.0.0 (2021-03-20) NEXT
 
 ## Chore
 
 - @capacitor/core@3.0.0-rc.0
 - @capacitor-community/sqlite@3.0.0-beta.7
 
-## 2.0.0-alpha.1 (2021-01-26) REFACTOR
+## 2.0.0-alpha.1 (2021-01-26) NEXT
 
 ## Chore